Interface changes (mostly size_t and number of elements in FieldList)

+++ src/PYB11/Boundary/ConstantBoundary.py
@@ -72,7 +72,7 @@ NodeList.
     # Properties
     nodeList = PYB11property("const NodeList<%(Dimension)s>&", "nodeList", doc="The NodeList this boundary applies to")
     nodeIndices = PYB11property("std::vector<size_t>", "nodeIndices", doc="The nodes this boundary is in control of")
-    numConstantNodes = PYB11property("int", "numConstantNodes", doc="The number of nodes we are controlling")
+    numConstantNodes = PYB11property("size_t", "numConstantNodes", doc="The number of nodes we are controlling")
     reflectOperator = PYB11property("const Tensor", "reflectOperator", doc="The tensor reflection transformation")
 
 #-------------------------------------------------------------------------------
